LayoutTests/imported/w3c/web-platform-tests/html/canvas/offscreen/text/2d.text.font.weight.worker-expected.txt


FAIL 2d assert_equals: ctx.font === 'italic 300 12px serif' (got italic 12px serif[string], expected italic 300 12px serif[string]) expected "italic 300 12px serif" but got "italic 12px serif"
